+++ apx_sdk/CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `is_verification_method` function.
 - Implemented `PartialEq` for `Hostname` type.
 - Added `parse_canonical` method to `Url` type.
+- Added `is_onion` and `is_i2p` methods to `apx_core::url::hostname` module.
 
 ### Changed